requirements.txt
# Note: PyTorch (torch and torchvision) should be installed separately 
# based on your CUDA version. Install from: https://pytorch.org/

# Core numerical computing
numpy==1.23.0

# Computer vision libraries
opencv-python==4.7.0.72
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.72
pillow==9.5.0
albumentations==1.3.0
imageio==2.26.0
scikit-image==0.20.0

# MMlab ecosystem
mmcv-full==1.6.0
mmengine==0.6.0

# Machine learning utilities
scipy==1.8.1
scikit-learn==1.2.2
pandas==1.3.5

# Neural network utilities
einops==0.6.0
timm==0.5.4

# Training and logging
tensorboard==2.12.0
lightning==2.0.1
tqdm==4.65.0

# Model optimization and deployment
thop==0.1.1-2209072238

# Utilities
addict==2.4.0
pyyaml==6.0
fire==0.5.0

# Code quality and testing
yapf==0.40.1
pytest==7.4.2

# Visualization and output
matplotlib==3.8.0
prettytable==3.9.0
terminaltables==3.1.10

# Build tools
setuptools==68.2.2
setuptools-scm==8.0.4

# Additional detection/segmentation
ultralytics==8.0.132

opt_einsum=3.4.0
ruamel.yaml==0.18.16


